Ensure that `kubernetes` client is always reset for all tests globally
