+++ tests/cli/conftest.py
@@ -4,7 +4,6 @@ import sys
 import click.testing
 import pytest
 
-import kopf
 from kopf.cli import main
 
 SCRIPT1 = """
@@ -50,16 +49,6 @@ def clean_modules_cache():
     for key in list(sys.modules.keys()):
         if key.startswith('package'):
             del sys.modules[key]
-
-
-@pytest.fixture(autouse=True)
-def clean_kubernetes_client():
-    try:
-        import kubernetes
-    except ImportError:
-        pass  # absent client is already "clean" (or not "dirty" at least).
-    else:
-        kubernetes.client.configuration.Configuration.set_default(None)
 
 
 @pytest.fixture()
