+++ tests/conftest.py
@@ -190,6 +190,16 @@ def login_mocks(mocker):
         )
     return LoginMocks(**kwargs)
 
+
+@pytest.fixture(autouse=True)
+def clean_kubernetes_client():
+    try:
+        import kubernetes
+    except ImportError:
+        pass  # absent client is already "clean" (or not "dirty" at least).
+    else:
+        kubernetes.client.configuration.Configuration.set_default(None)
+
 #
 # Simulating that Kubernetes client libraries are not installed.
 #
